Refactor: move parts and relations to own file
